Deprecate digital_ocean.py in favor of digital_ocean_v2.py
The removes support for the original digital_ocean.py driver in salt-cloud
and replaces the original digital_ocean.py driver with the digital_ocean_v2.py
driver. The digital_ocean_v2.py driver is renamed to just digital_ocean.py.
